# orders-api/openapi.yaml
openapi: 3.0.3
info:
  title: Orders API
  version: 1.0.0
  description: Gestión de productos y órdenes con validación de stock e idempotencia.
servers:
  - url: http://localhost:3002
security:
  - bearerAuth: []
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /products:
    post:
      summary: Crear producto
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    get:
      summary: Listar productos
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: cursor
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  nextCursor:
                    type: integer
  /products/{id}:
    get:
      summary: Obtener producto
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Producto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
    patch:
      summary: Actualizar producto
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
  /orders:
    post:
      summary: Crear orden (valida cliente y stock)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Orden creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
    get:
      summary: Listar órdenes
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [CREATED, CONFIRMED, CANCELED]
        - in: query
          name: from
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          schema:
            type: string
            format: date-time
        - in: query
          name: cursor
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Lista paginada
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Order'
                  nextCursor:
                    type: integer
  /orders/{id}:
    get:
      summary: Obtener orden con items
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Orden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /orders/{id}/confirm:
    post:
      summary: Confirmar orden (idempotente)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
        - in: header
          name: X-Idempotency-Key
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confirmada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
  /orders/{id}/cancel:
    post:
      summary: Cancelar orden (restaura stock)
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Cancelada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
  schemas:
    Product:
      type: object
      properties:
        id:
          type: integer
        sku:
          type: string
        name:
          type: string
        price_cents:
          type: integer
        stock:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductCreate:
      type: object
      required: [sku, name, price_cents, stock]
      properties:
        sku:
          type: string
        name:
          type: string
        price_cents:
          type: integer
        stock:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        price_cents:
          type: integer
        stock:
          type: integer
    ProductResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Product'
    OrderItem:
      type: object
      properties:
        product_id:
          type: integer
        qty:
          type: integer
        unit_price_cents:
          type: integer
        subtotal_cents:
          type: integer
    Order:
      type: object
      properties:
        id:
          type: integer
        customer_id:
          type: integer
        status:
          type: string
        total_cents:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderCreate:
      type: object
      required: [customer_id, items]
      properties:
        customer_id:
          type: integer
        items:
          type: array
          items:
            type: object
            required: [product_id, qty]
            properties:
              product_id:
                type: integer
              qty:
                type: integer
    OrderResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Order'